Added more props still not done.
All props and chests and keys (story) must be completed by the end of tomorrow. Also the book needs finishing.

Once all of that is done then add the enemy (this should take 1 day (maybe 2 at the most)). The enemy will still need the stun and attack method written. Also if time is there make it so the axe can be used to have a longer reach to stun.