Added branch for AI testing
I am having issues with the AI all of suddenly just not being able to walk up the stairs when I haven't even re-baked the Nav Mesh Surface; all I've done is add Nav Mesh Obstacles and ticked the Carve box too. (This is quite good because it does actually make the AI walk around to the nearest and possible way to get to the player. Though just be doing these changes the AI can no longer move up the stairs.