Added in the last rooms; just gotta add in the probs and keys (maybe see if random chests are simple to do?)
Also in the AI Controller script just need to add the OnTriggerStay open the door (that should help fix the possibility of it getting stuck inside of the door. Another thing to add is when the AI opens and closes door I should play the sound effect for that. (Though I seriously need to figure out how to make it 3D or world sound that way it doesn't sound right next to you.

Lastly I just need sounds for the AI movement and the bookshelf falling over.